Patch C
Added the beta wing palace rabbit
Added ice bullies (I think there's only one)
Added toads
Fixed boos
Added a bunch of object specific stars (like wigglers and piranha plants)
Attempted to resolve a conflict between a freestanding red star and a boss reward, though I'm unsure if the freestanding star is actually accessible, as I have not found it yet.

Fixed interpreter-specific crash with the suns (this crash does not happen when using the recompiler).

Added a tweak to disable that music box effect at night. Like the other tweaks this is not present in the patch.